+++ packages/core/src/event-bus/index.ts
@@ -1,24 +1,34 @@
-import { SocketEventHandler, SocketListener } from './socket';
+import { SAEventTarget } from '../common';
 
 export * from './module';
 export * from './socket';
 
 export class SAEventBus {
-    callbacks: CallBack[] = [];
+    private cleanFn: CallBack[] = [];
 
-    sac(event: string, callback: CallBack) {
-        this.callbacks.push(callback);
+    hook<TEvent extends string>(...args: Parameters<typeof SAEventTarget.on<TEvent>>) {
+        this.cleanFn.push(() => {
+            SAEventTarget.off(...args);
+        });
+        return SAEventTarget.on(...args);
     }
 
-    game(event: string, callback: CallBack) {
-        this.callbacks.push(callback);
+    socket<This>(...args: Parameters<typeof SocketConnection.addCmdListener<This>>) {
+        this.cleanFn.push(() => {
+            SocketConnection.removeCmdListener(...args);
+        });
+        return SocketConnection.addCmdListener(...args);
     }
 
-    socket<TCmd extends number>(handler: SocketEventHandler<TCmd>) {
-        // this.callbacks.push(callback);
+    egret<T extends egret.Event>(event: string, callback: CallBack<T>) {
+        this.cleanFn.push(() => {
+            EventManager.removeEventListener(event, callback, undefined);
+        });
+        return EventManager.addEventListener(event, callback, undefined);
     }
 
     unmount() {
-        this.callbacks = [];
+        this.cleanFn.forEach((fn) => fn());
+        this.cleanFn = [];
     }
 }