+++ packages/core/src/pet-helper/internal.ts
@@ -1,9 +1,10 @@
+import { SAEventTarget } from '../common';
+import { Hook } from '../constant';
 import { SocketListener } from '../event-bus';
 import { PetDataManger, ProxyPet } from './ProxyPet';
 
 export default () => {
-    SocketListener.subscribe(
-        CommandID.GET_PET_INFO_BY_ONCE, (data) => {
+    SocketListener.subscribe(CommandID.GET_PET_INFO_BY_ONCE, (data) => {
         const bytes = new egret.ByteArray(data);
         let size = bytes.readUnsignedInt();
         const r1 = [];
@@ -35,7 +36,19 @@ export default () => {
     SocketListener.subscribe(CommandID.DEL_LOVE_PET);
     SocketListener.subscribe(CommandID.PET_CURE);
 
-    new Promise<void>((r) => PetStorage2015InfoManager.getMiniInfo(r)).then(() => {
-        PetDataManger.init();
+    PetDataManger.init();
+
+    PetDataManger.bag.deactivate = new Proxy(PetDataManger.bag.deactivate, {
+        apply: (target, thisArg, argArray) => {
+            SAEventTarget.emit(Hook.PetBag.deactivate);
+            return Reflect.apply(target, thisArg, argArray);
+        },
+    });
+
+    PetDataManger.bag.update = new Proxy(PetDataManger.bag.update, {
+        apply: (target, thisArg, argArray) => {
+            SAEventTarget.emit(Hook.PetBag.update);
+            return Reflect.apply(target, thisArg, argArray);
+        },
     });
 };